Working super basic version of OLED support
This has a ton of cleanup still needed, but it does bring the OLED online for the new ESP32 board being tested

Next up is some additional testing to confirm if we actually need pin16 high and a ton of cleanup.  Part of the cleanup should include sticking the OLED code into different modules and choosing what we load based on the actual board.

Note - The code base will be targeted towards specific hardware configs, but we want to keep some open source flexibility in the design